dashboard/requirements.txt
# Dashboard Python Dependencies

# Garmin Connect API
garminconnect>=0.2.8
garth>=0.4.23

# InfluxDB Client
influxdb-client>=1.44.0

# Environment management
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# JSON Schema validation
jsonschema>=4.17.3